pyproject.toml
[tool.poetry]
name = "olpxek-bot"
version = "0.1.0"
description = ""
authors = ["Chanwoong Kim <me@chanwoong.kim>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
"discord.py" = "^2.3.2"
RestrictedPython = "^5.0"
playwright = "^1.18.1"
cachetools = "^5.3.1"
asyncache = "^0.3.1"
bs4 = "^0.0.1"
juga = {git = "https://github.com/kexplo/juga", rev = "f8c56b6"}
typer = "^0.9.0"
sentry-sdk = "^1.0.0"
omegaconf = "^2.0.6"
matplotlib = "^3.3.4"
pandas = "^1.2.3"
mplfinance = "^0.12.7-alpha.10"
rich = "^13.6.0"
olpxek-bot-private-cogs = {path = "../olpxek-bot-private-cogs", develop = true}
youtube-transcript-api = "^0.6.2"

[tool.poetry.dev-dependencies]
ptpython = "^3.0.3"
mypy = "*"
black = "*"
flake8-import-order = "^0.18.2"
pep8-naming = "^0.13.3"
flake8-pytest-style = "^1.2.3"
flake8-logging-format = "^0.9.0"
flake8-comprehensions = "^3.14.0"
flake8-builtins = "^2.1.0"
flake8-bugbear = "^23.7.10"
flake8-bandit = "^4.1.1"
pytest = "^7.4.0"
aioresponses = "^0.7.4"
pytest-asyncio = "^0.21.1"

[tool.poetry.scripts]
olpxek_bot = 'olpxek_bot.__main__:app'

[tool.poetry.group.dev.dependencies]
ruff = "^0.0.290"

[tool.isort]
profile = "google"

[tool.black]
line-length = 120

[tool.ruff]
line-length = 120

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
